+++ web/apps/photos/src/services/upload/uploadManager.ts
@@ -5,7 +5,7 @@ import { lowercaseExtension, nameAndExtension } from "@/base/file-name";
 import log from "@/base/log";
 import type { Electron } from "@/base/types/ipc";
 import { ComlinkWorker } from "@/base/worker/comlink-worker";
-import { shouldDisableCFUploadProxy } from "@/gallery/upload";
+import { shouldDisableCFUploadProxy } from "@/gallery/services/upload";
 import type { Collection } from "@/media/collection";
 import { EncryptedEnteFile, EnteFile } from "@/media/file";
 import type { ParsedMetadata } from "@/media/file-metadata";
